Add financial entities and dashboard updates

+++ Core/Application/DTO/IncomePlannningDTO.cs
@@ -3,8 +3,11 @@ namespace Application.DTO.IncomePlanning{
 
         public int BudgetId {get;set;}
         public int IncomeTypeId {get;set;}
-        public decimal Amount{get;set;}
+        public double Amount{get;set;}
         public string Description {get;set;}
+        // public DateTime CreatedAt {get;set;}
+        // public int UserAdded{get;set;}
+        // public DateTime UpdatedAt{get;set;}
 
     }
 }
